+++ example/lib/main.dart
@@ -5,7 +5,7 @@ void main() => runApp(MyApp());
 
 class MyApp extends StatefulWidget {
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
@@ -16,6 +16,7 @@ class _MyAppState extends State<MyApp> {
   String bearing = 'waiting...';
   String speed = 'waiting...';
   String time = 'waiting...';
+  bool? serviceRunning;
 
   @override
   void initState() {
@@ -32,13 +33,14 @@ class _MyAppState extends State<MyApp> {
         body: Center(
           child: ListView(
             children: <Widget>[
-              locationData('Latitude: ' + latitude),
-              locationData('Longitude: ' + longitude),
-              locationData('Altitude: ' + altitude),
-              locationData('Accuracy: ' + accuracy),
-              locationData('Bearing: ' + bearing),
-              locationData('Speed: ' + speed),
-              locationData('Time: ' + time),
+              locationData('Latitude: $latitude'),
+              locationData('Longitude: $longitude'),
+              locationData('Altitude: $altitude'),
+              locationData('Accuracy: $accuracy'),
+              locationData('Bearing: $bearing'),
+              locationData('Speed: $speed'),
+              locationData('Time: $time'),
+              locationData('IsServiceRunning: $serviceRunning'),
               ElevatedButton(
                   onPressed: () async {
                     await BackgroundLocation.setAndroidNotification(
@@ -69,6 +71,7 @@ class _MyAppState extends State<MyApp> {
                         Bearing:  $bearing
                         Speed: $speed
                         Time: $time
+                        IsServiceRunning: $serviceRunning
                       ''');
                     });
                   },
@@ -78,6 +81,16 @@ class _MyAppState extends State<MyApp> {
                     BackgroundLocation.stopLocationService();
                   },
                   child: Text('Stop Location Service')),
+              ElevatedButton(
+                  onPressed: () {
+                    BackgroundLocation.isServiceRunning().then((value) {
+                      setState(() {
+                        serviceRunning = value;
+                      });
+                      print('Is Running: $value');
+                    });
+                  },
+                  child: Text('Check service')),
               ElevatedButton(
                   onPressed: () {
                     getCurrentLocation();
@@ -103,7 +116,7 @@ class _MyAppState extends State<MyApp> {
 
   void getCurrentLocation() {
     BackgroundLocation().getCurrentLocation().then((location) {
-      print('This is current Location ' + location.toMap().toString());
+      print('This is current Location ${location.toMap()}');
     });
   }