+++ lib/background_location.dart
@@ -13,33 +13,35 @@ class BackgroundLocation {
       MethodChannel('background.location');
 
   /// Stop receiving location updates
-  static stopLocationService() async {
+  static Future<dynamic> stopLocationService() async {
     return await _channel.invokeMethod('stop_location_service');
   }
 
+  /// Check if the location update service is running
+  static Future<bool> isServiceRunning() async {
+    var result = await _channel.invokeMethod('is_service_running');
+    return result == true;
+  }
+
   /// Start receiving location updated
-  static startLocationService({double distanceFilter = 0.0, bool forceAndroidLocationManager = false}) async {
+  static Future<dynamic> startLocationService({double distanceFilter = 0.0, bool forceAndroidLocationManager = false}) async {
     return await _channel.invokeMethod('start_location_service',
         <String, dynamic>{'distance_filter': distanceFilter, 'force_location_manager': forceAndroidLocationManager});
   }
 
-  static setAndroidNotification(
+  static Future<dynamic> setAndroidNotification(
       {String? title, String? message, String? icon}) async {
     if (Platform.isAndroid) {
       return await _channel.invokeMethod('set_android_notification',
           <String, dynamic>{'title': title, 'message': message, 'icon': icon});
-    } else {
-      //return Promise.resolve();
     }
   }
 
-  static setAndroidConfiguration(int interval) async {
+  static Future<dynamic> setAndroidConfiguration(int interval) async {
     if (Platform.isAndroid) {
       return await _channel.invokeMethod('set_configuration', <String, dynamic>{
         'interval': interval.toString(),
       });
-    } else {
-      //return Promise.resolve();
     }
   }
 
@@ -47,16 +49,18 @@ class BackgroundLocation {
   Future<Location> getCurrentLocation() async {
     var completer = Completer<Location>();
 
-    var _location = Location();
-    await getLocationUpdates((location) {
-      _location.latitude = location.latitude;
-      _location.longitude = location.longitude;
-      _location.accuracy = location.accuracy;
-      _location.altitude = location.altitude;
-      _location.bearing = location.bearing;
-      _location.speed = location.speed;
-      _location.time = location.time;
-      completer.complete(_location);
+    getLocationUpdates((location) {
+      var loc = Location(
+        latitude: location.latitude,
+        longitude: location.longitude,
+        accuracy: location.accuracy,
+        altitude: location.altitude,
+        bearing: location.bearing,
+        speed: location.speed,
+        time: location.time,
+        isMock: location.isMock,
+      );
+      completer.complete(loc);
     });
 
     return completer.future;
@@ -66,7 +70,7 @@ class BackgroundLocation {
 
   /// Register a function to recive location updates as long as the location
   /// service has started
-  static getLocationUpdates(Function(Location) location) {
+  static void getLocationUpdates(Function(Location) location) {
     // add a handler on the channel to recive updates from the native classes
     _channel.setMethodCallHandler((MethodCall methodCall) async {
       if (methodCall.method == 'location') {
@@ -109,7 +113,7 @@ class Location {
       @required this.time,
       @required this.isMock});
 
-  toMap() {
+  Map<String, dynamic> toMap() {
     var obj = {
       'latitude': latitude,
       'longitude': longitude,