+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+### 0.13.2
+
+- Fixed issues with android namespaces
+- resolve flutter analyze lint errors and update flutter lint
+- Clarify permissions requirements earlier in the documentation [#226](https://github.com/Almoullim/background_location/pull/#226)
+
+### 0.13.1
+
+- Fix License Typo [#201](https://github.com/Almoullim/background_location/pull/#201)
+- Fix API issues stemming from Android 13/14 changes to location services [#210](https://github.com/Almoullim/background_location/pull/#210)
+
+### 0.13.0
+
+- Android upgrade and location running indicator [#196](https://github.com/Almoullim/background_location/pull/196)
+
 ### 0.12.0
 - Android: Raise notification importance [#163](https://github.com/Almoullim/background_location/issues/167)